separate SQLiteCache::Impl header from implementation file
